Refactored general purpose functions from stage_dem.py into a new geo_utils.py module
Added support for specifying the source S3 bucket as a command-line argument to stage_dem.py
