+++ tools/stage_dem.py
@@ -7,23 +7,23 @@ import os
 import backoff
 
 import boto3
-import mgrs
-import numpy as np
-import pyproj
-import shapely.ops
 import shapely.wkt
+
+from osgeo import gdal
+from shapely.geometry import Polygon, box
 
 from commons.logger import logger
 from commons.logger import LogLevels
-
-from osgeo import gdal, osr
-from shapely.geometry import LinearRing, Point, Polygon, box
+from util.geo_util import (check_dateline,
+                           epsg_from_polygon,
+                           polygon_from_mgrs_tile,
+                           transform_polygon_coords_to_epsg)
 
 # Enable exceptions
 gdal.UseExceptions()
 
 S3_DEM_BUCKET = "opera-dem"
-"""Name of the S3 bucket containing the full DEM's to crop from"""
+"""Name of the default S3 bucket containing the global DEM to crop from"""
 
 
 def get_parser():
@@ -40,14 +40,18 @@ def get_parser():
                              'to determine overlap between provided DEM, and '
                              'DEM to be downloaded based on the MGRS tile code '
                              'or bounding box.')
-    parser.add_argument('-m', '--margin', type=int, action='store',
-                        default=5, help='Margin for DEM bounding box in km.')
+    parser.add_argument('-s', '--s3-bucket', type=str, action='store',
+                        default=S3_DEM_BUCKET, dest='s3_bucket',
+                        help='Name of the S3 bucket containing the global DEM '
+                             'to extract from.')
+    parser.add_argument('-t', '--tile-code', type=str, default=None,
+                        help='MGRS tile code identifier for the DEM region')
     parser.add_argument('-b', '--bbox', type=float, action='store',
                         dest='bbox', default=None, nargs='+',
                         help='Spatial bounding box of the DEM region in '
                              'latitude/longitude (WSEN, decimal degrees)')
-    parser.add_argument('-t', '--tile-code', type=str, default=None,
-                        help='MGRS tile code identifier for the DEM region')
+    parser.add_argument('-m', '--margin', type=int, action='store',
+                        default=5, help='Margin for DEM bounding box in km.')
     parser.add_argument("--log-level",
                         type=lambda log_level: LogLevels[log_level].value,
                         choices=LogLevels.list(),
@@ -55,48 +59,6 @@ def get_parser():
                         help="Specify a logging verbosity level.")
 
     return parser
-
-
-def check_dateline(poly):
-    """
-    Split `poly` if it crosses the dateline.
-
-    Parameters
-    ----------
-    poly : shapely.geometry.Polygon
-        Input polygon.
-
-    Returns
-    -------
-    polys : list of shapely.geometry.Polygon
-        A list containing: the input polygon if it didn't cross the dateline, or
-        two polygons otherwise (one on either side of the dateline).
-
-    """
-    x_min, _, x_max, _ = poly.bounds
-
-    # Check dateline crossing
-    if (x_max - x_min) > 180.0:
-        dateline = shapely.wkt.loads('LINESTRING( 180.0 -90.0, 180.0 90.0)')
-
-        # build new polygon with all longitudes between 0 and 360
-        x, y = poly.exterior.coords.xy
-        new_x = (k + (k <= 0.) * 360 for k in x)
-        new_ring = LinearRing(zip(new_x, y))
-
-        # Split input polygon
-        # (https://gis.stackexchange.com/questions/232771/splitting-polygon-by-linestring-in-geodjango_)
-        merged_lines = shapely.ops.linemerge([dateline, new_ring])
-        border_lines = shapely.ops.unary_union(merged_lines)
-        decomp = shapely.ops.polygonize(border_lines)
-
-        polys = list(decomp)
-        assert (len(polys) == 2)
-    else:
-        # If dateline is not crossed, treat input poly as list
-        polys = [poly]
-
-    return polys
 
 
 def determine_polygon(tile_code, bbox=None):
@@ -119,146 +81,22 @@ def determine_polygon(tile_code, bbox=None):
         the ground.
 
     """
-    if bbox is not None:
+    if bbox:
         logger.info('Determining polygon from bounding box')
         poly = box(bbox[0], bbox[1], bbox[2], bbox[3])
     else:
         logger.info(f'Determining polygon from MGRS tile code {tile_code}')
-        poly = get_polygon_from_mgrs(tile_code)
+        poly = polygon_from_mgrs_tile(tile_code)
 
     logger.debug(f'Derived polygon {str(poly)}')
 
     return poly
 
 
-def point2epsg(lon, lat):
-    """
-    Return an EPSG code based on the provided lat/lon point.
-
-    Parameters
-    ----------
-    lat: float
-        Latitude coordinate of the point
-    lon: float
-        Longitude coordinate of the point
-
-    Returns
-    -------
-    EPSG code corresponding to the point lat/lon coordinates.
-
-    Raises
-    ------
-    ValueError
-        If the EPSG code cannot be determined from the provided lat/lon.
-
-    """
-    if lon >= 180.0:
-        lon = lon - 360.0
-    if lat >= 75.0:
-        return 3413
-    elif lat <= -75.0:
-        return 3031
-    elif lat > 0:
-        return 32601 + int(np.round((lon + 177) / 6.0))
-    elif lat < 0:
-        return 32701 + int(np.round((lon + 177) / 6.0))
-    else:
-        raise ValueError(f'Could not determine projection for {lat},{lon}')
-
-
-def get_polygon_from_mgrs(tile_code):
-    """
-    Create a polygon (EPSG:4326) from the lat/lon coordinates corresponding to
-    a MGRS tile bounding box.
-
-    Parameters
-    -----------
-    tile_code : str
-        MGRS tile code corresponding to the polygon to derive.
-
-    Returns
-    -------
-    poly: shapely.Geometry.Polygon
-        Bounding polygon corresponding to the provided MGRS tile code.
-
-    """
-    mgrs_obj = mgrs.MGRS()
-
-    geod = pyproj.Geod(ellps='WGS84')
-
-    if tile_code.startswith('T'):
-        tile_code = tile_code[1:]
-
-    lat_min, lon_min = mgrs_obj.toLatLon(tile_code, inDegrees=True)
-    x_var = geod.line_length([lon_min, lon_min], [lat_min, lat_min + 1])
-    y_var = geod.line_length([lon_min, lon_min + 1], [lat_min, lat_min])
-
-    mgrs_tile_edge_size = 109.8 * 1000
-
-    lat_max = lat_min + (mgrs_tile_edge_size / x_var)
-    lon_max = lon_min + (mgrs_tile_edge_size / y_var)
-
-    coords = list(map(round, [lon_min, lat_min, lon_max, lat_max]))
-
-    poly = box(*coords)
-
-    return poly
-
-
-def determine_projection(polys):
-    """
-    Determine EPSG code for each polygon in polys.
-
-    EPSG is computed for a regular list of points. EPSG is assigned based on a
-    majority criteria.
-
-    Parameters
-    -----------
-    polys: list of shapely.Geometry.Polygon
-        List of shapely Polygons
-
-    Returns
-    -------
-    epsgs: list of int
-        List of EPSG codes corresponding to elements in polys
-
-    """
-    logger.info("Determining EPSG code(s) for region polygon(s)")
-
-    epsgs = []
-
-    # Make a regular grid based on polys min/max latitude longitude
-    for p in polys:
-        x_min, y_min, x_max, y_max = p.bounds
-        xx, yy = np.meshgrid(np.linspace(x_min, x_max, 250),
-                             np.linspace(y_min, y_max, 250))
-        x = xx.flatten()
-        y = yy.flatten()
-
-        # Query to determine the zone
-        zones = []
-        for lx, ly in zip(x, y):
-            # Create a point with grid coordinates
-            pp = Point(lx, ly)
-
-            # If Point is in polys, compute EPSG
-            if pp.within(p):
-                zones.append(point2epsg(lx, ly))
-
-        # Count different EPSGs
-        vals, counts = np.unique(zones, return_counts=True)
-
-        # Get the EPSG for Polys
-        epsgs.append(vals[np.argmax(counts)])
-
-    logger.debug(f'Derived the following EPSG codes: {epsgs}')
-    return epsgs
-
-
 @backoff.on_exception(backoff.expo, Exception, max_tries=8, max_value=32)
 def translate_dem(vrt_filename, output_path, x_min, x_max, y_min, y_max):
     """
-    Translate a DEM from the opera-dem bucket.
+    Translate a DEM from S3 to a region matching the provided boundaries.
 
     Notes
     -----
@@ -291,9 +129,9 @@ def translate_dem(vrt_filename, output_path, x_min, x_max, y_min, y_max):
     )
 
 
-def download_dem(polys, epsgs, margin, outfile):
+def download_dem(polys, epsgs, dem_bucket, margin, outfile):
     """
-    Download a DEM from the opera-dem bucket.
+    Download a DEM from the specified S3 bucket.
 
     Parameters:
     ----------
@@ -301,6 +139,8 @@ def download_dem(polys, epsgs, margin, outfile):
         List of shapely polygons.
     epsgs: list of str
         List of EPSG codes corresponding to polys.
+    dem_bucket : str
+        Name of the S3 bucket containing the global DEM to download from.
     margin: float
         Buffer margin (in km) applied for DEM download.
     outfile:
@@ -309,14 +149,14 @@ def download_dem(polys, epsgs, margin, outfile):
     """
     if 3031 in epsgs:
         epsgs = [3031] * len(epsgs)
-        polys = transform_polygon_coords(polys, epsgs)
+        polys = transform_polygon_coords_to_epsg(polys, epsgs)
 
         # Need one EPSG as in polar stereo we have one big polygon
         epsgs = [3031]
         margin = margin * 1000
     elif 3413 in epsgs:
         epsgs = [3413] * len(epsgs)
-        polys = transform_polygon_coords(polys, epsgs)
+        polys = transform_polygon_coords_to_epsg(polys, epsgs)
 
         # Need one EPSG as in polar stereo we have one big polygon
         epsgs = [3413]
@@ -333,7 +173,7 @@ def download_dem(polys, epsgs, margin, outfile):
     dem_list = []
 
     for idx, (epsg, poly) in enumerate(zip(epsgs, polys)):
-        vrt_filename = f'/vsis3/{S3_DEM_BUCKET}/EPSG{epsg}/EPSG{epsg}.vrt'
+        vrt_filename = f'/vsis3/{dem_bucket}/EPSG{epsg}/EPSG{epsg}.vrt'
         poly = poly.buffer(margin)
         output_path = f'{file_prefix}_{idx}.tif'
         dem_list.append(output_path)
@@ -342,58 +182,6 @@ def download_dem(polys, epsgs, margin, outfile):
 
     # Build vrt with downloaded DEMs
     gdal.BuildVRT(outfile, dem_list)
-
-
-def transform_polygon_coords(polys, epsgs):
-    """
-    Transform coordinates of polys (list of polygons) to target epsgs (list of
-    EPSG codes).
-
-    Parameters
-    ----------
-    polys: list of shapely.Geometry.Polygon
-        List of shapely polygons
-    epsgs: list of str
-        List of EPSG codes corresponding to elements in polys
-
-    Returns
-    -------
-    poly : list of shapely.Geometry.Polygon
-         A list containing a single polygon which spans the extent of all
-         transformed polygons.
-
-    """
-    # Assert validity of inputs
-    assert(len(polys) == len(epsgs))
-
-    # Transform each point of the perimeter in target EPSG coordinates
-    llh = osr.SpatialReference()
-    llh.ImportFromEPSG(4326)
-    tgt = osr.SpatialReference()
-
-    x_min, y_min, x_max, y_max = [], [], [], []
-    tgt_x, tgt_y = [], []
-
-    for poly, epsg in zip(polys, epsgs):
-        x, y = poly.exterior.coords.xy
-        tgt.ImportFromEPSG(int(epsg))
-        trans = osr.CoordinateTransformation(llh, tgt)
-
-        for lx, ly in zip(x, y):
-            dummy_x, dummy_y, dummy_z = trans.TransformPoint(ly, lx, 0)
-            tgt_x.append(dummy_x)
-            tgt_y.append(dummy_y)
-
-        x_min.append(min(tgt_x))
-        y_min.append(min(tgt_y))
-        x_max.append(max(tgt_x))
-        y_max.append(max(tgt_y))
-
-    # return a polygon
-    poly = [Polygon([(min(x_min), min(y_min)), (min(x_min), max(y_max)),
-                     (max(x_max), max(y_max)), (max(x_max), min(y_min))])]
-
-    return poly
 
 
 def check_dem_overlap(dem_filepath, polys):
@@ -428,7 +216,7 @@ def check_dem_overlap(dem_filepath, polys):
     epsg = [DEM.get_epsg()] * len(polys)
 
     if DEM.get_epsg() != 4326:
-        polys = transform_polygon_coords(polys, epsg)
+        polys = transform_polygon_coords_to_epsg(polys, epsg)
 
     perc_area = 0
     for poly in polys:
@@ -437,9 +225,14 @@ def check_dem_overlap(dem_filepath, polys):
     return perc_area
 
 
-def check_aws_connection():
+def check_aws_connection(dem_bucket):
     """
-    Check connection to the AWS s3://opera-dem bucket.
+    Check connection to the provided S3 bucket.
+
+    Parameters
+    ----------
+    dem_bucket : str
+        Name of the bucket to use with the connection test.
 
     Raises
     ------
@@ -447,15 +240,15 @@ def check_aws_connection():
        If no connection can be established.
 
     """
-    logger.info(f'Checking connection to AWS S3 {S3_DEM_BUCKET} bucket.')
     s3 = boto3.resource('s3')
-    obj = s3.Object(S3_DEM_BUCKET, 'EPSG4326/EPSG4326.vrt')
+    obj = s3.Object(dem_bucket, 'EPSG4326/EPSG4326.vrt')
 
     try:
+        logger.info(f'Attempting test read of s3://{obj.bucket_name}/{obj.key}')
         obj.get()['Body'].read()
         logger.info('Connection test successful.')
     except Exception:
-        errmsg = (f'No access to the {S3_DEM_BUCKET} s3 bucket. '
+        errmsg = (f'No access to the {dem_bucket} s3 bucket. '
                   f'Check your AWS credentials and re-run the code.')
         raise RuntimeError(errmsg)
 
@@ -485,6 +278,11 @@ def main(opts):
         err_msg = "DEM output filename extension is not .vrt"
         raise ValueError(err_msg)
 
+    # Check if we were provided an explicit "None" for the s3_bucket,
+    # which can occur when arguments are set up by a chimera precondition function
+    if not opts.s3_bucket:
+        opts.s3_bucket = S3_DEM_BUCKET
+
     # Determine polygon based on MGRS info or bbox
     poly = determine_polygon(opts.tile_code, opts.bbox)
 
@@ -505,14 +303,20 @@ def main(opts):
             logger.warning('Unable to import from isce3 package, cannot determine '
                            'DEM overlap.')
 
-    # Check connection to AWS s3 opera-dem bucket
-    check_aws_connection()
+    # Check connection to the S3 bucket
+    logger.info(f'Checking connection to AWS S3 {opts.s3_bucket} bucket.')
+
+    check_aws_connection(opts.s3_bucket)
 
     # Determine EPSG code
-    epsgs = determine_projection(polys)
+    logger.info("Determining EPSG code(s) for region polygon(s)")
+
+    epsgs = epsg_from_polygon(polys)
+
+    logger.debug(f'Derived the following EPSG codes: {epsgs}')
 
     # Download DEM
-    download_dem(polys, epsgs, opts.margin, opts.outfile)
+    download_dem(polys, epsgs, opts.s3_bucket, opts.margin, opts.outfile)
 
     logger.info(f'Done, DEM stored locally to {opts.outfile}')
